requirements.txt
fastapi[all]
pyjwt[crypto]